+++ relayer-cli/src/commands/query/client.rs
@@ -19,19 +19,25 @@ use crate::conclude::Output;
 use crate::error::{Error, Kind};
 use crate::prelude::*;
 
+// TODO: refactor commands: why is chain_id an `Option`? simpler to give it `ChainId` type.
+//  see the tx/client.rs or tx/packet.rs for simpler approaches.
+
 /// Query client state command
 #[derive(Clone, Command, Debug, Options)]
 pub struct QueryClientStateCmd {
-    #[options(free, help = "identifier of the chain to query")]
-    chain_id: Option<ChainId>,
+    #[options(free, required, help = "identifier of the chain to query")]
+    chain_id: ChainId,
 
-    #[options(free, help = "identifier of the client to query")]
+    #[options(free, required, help = "identifier of the client to query")]
     client_id: Option<String>,
 
     #[options(help = "the chain height which this query should reflect", short = "h")]
     height: Option<u64>,
 
-    #[options(help = "whether proof is required", short = "p")]
+    #[options(
+        help = "whether proof is required; default: false (no proof)",
+        short = "p"
+    )]
     proof: Option<bool>,
 }
 
@@ -95,16 +101,24 @@ impl Runnable for QueryClientStateCmd {
 /// Query client consensus command
 #[derive(Clone, Command, Debug, Options)]
 pub struct QueryClientConsensusCmd {
-    #[options(free, help = "identifier of the chain to query")]
-    chain_id: Option<ChainId>,
+    #[options(free, required, help = "identifier of the chain to query")]
+    chain_id: ChainId,
 
-    #[options(free, help = "identifier of the client to query")]
+    #[options(free, required, help = "identifier of the client to query")]
     client_id: Option<String>,
 
-    #[options(free, help = "epoch of the client's consensus state to query")]
+    #[options(
+        free,
+        required,
+        help = "epoch of the client's consensus state to query"
+    )]
     consensus_epoch: Option<u64>,
 
-    #[options(free, help = "height of the client's consensus state to query")]
+    #[options(
+        free,
+        required,
+        help = "height of the client's consensus state to query"
+    )]
     consensus_height: Option<u64>,
 
     #[options(help = "the chain height which this query should reflect", short = "h")]
@@ -192,16 +206,13 @@ impl Runnable for QueryClientConsensusCmd {
 }
 
 fn validate_common_options(
-    chain_id: &Option<ChainId>,
+    chain_id: &ChainId,
     client_id: &Option<String>,
     config: &Config,
 ) -> Result<(ChainConfig, ClientId), String> {
-    let chain_id = chain_id
-        .clone()
-        .ok_or_else(|| "missing chain parameter".to_string())?;
     let chain_config = config
         .find_chain(&chain_id)
-        .ok_or_else(|| "missing chain in configuration".to_string())?;
+        .ok_or_else(|| format!("chain '{}' not found in configuration file", chain_id))?;
 
     let client_id = client_id
         .as_ref()
@@ -215,10 +226,10 @@ fn validate_common_options(
 /// Query client connections command
 #[derive(Clone, Command, Debug, Options)]
 pub struct QueryClientConnectionsCmd {
-    #[options(free, help = "identifier of the chain to query")]
-    chain_id: Option<ChainId>,
+    #[options(free, required, help = "identifier of the chain to query")]
+    chain_id: ChainId,
 
-    #[options(free, help = "identifier of the client to query")]
+    #[options(free, required, help = "identifier of the client to query")]
     client_id: Option<String>,
 
     #[options(help = "the chain height which this query should reflect", short = "h")]
@@ -236,13 +247,9 @@ impl QueryClientConnectionsCmd {
         &self,
         config: &Config,
     ) -> Result<(ChainConfig, QueryClientConnectionsOptions), String> {
-        let chain_id = self
-            .chain_id
-            .clone()
-            .ok_or_else(|| "missing chain identifier".to_string())?;
         let chain_config = config
-            .find_chain(&chain_id)
-            .ok_or_else(|| "missing chain configuration".to_string())?;
+            .find_chain(&self.chain_id)
+            .ok_or_else(|| format!("chain '{}' not found in configuration file", self.chain_id))?;
 
         let client_id = self
             .client_id
@@ -300,7 +307,7 @@ mod tests {
     #[test]
     fn parse_query_state_parameters() {
         let default_params = QueryClientStateCmd {
-            chain_id: Some("ibc-0".to_string().parse().unwrap()),
+            chain_id: "ibc-0".to_string().parse().unwrap(),
             client_id: Some("ibconeclient".to_string().parse().unwrap()),
             height: None,
             proof: None,
@@ -319,17 +326,9 @@ mod tests {
                 want_pass: true,
             },
             Test {
-                name: "No chain specified".to_string(),
-                params: QueryClientStateCmd {
-                    chain_id: None,
-                    ..default_params.clone()
-                },
-                want_pass: false,
-            },
-            Test {
                 name: "Chain not configured".to_string(),
                 params: QueryClientStateCmd {
-                    chain_id: Some("notibc0oribc1".to_string().parse().unwrap()),
+                    chain_id: "notibc0oribc1".to_string().parse().unwrap(),
                     ..default_params.clone()
                 },
                 want_pass: false,
@@ -386,7 +385,7 @@ mod tests {
     #[test]
     fn parse_query_client_connections_parameters() {
         let default_params = QueryClientConnectionsCmd {
-            chain_id: Some("ibc-0".to_string().parse().unwrap()),
+            chain_id: "ibc-0".to_string().parse().unwrap(),
             client_id: Some("clientidone".to_string().parse().unwrap()),
             height: Some(4),
         };
@@ -404,17 +403,9 @@ mod tests {
                 want_pass: true,
             },
             Test {
-                name: "No chain specified".to_string(),
-                params: QueryClientConnectionsCmd {
-                    chain_id: None,
-                    ..default_params.clone()
-                },
-                want_pass: false,
-            },
-            Test {
                 name: "Chain not configured".to_string(),
                 params: QueryClientConnectionsCmd {
-                    chain_id: Some("notibc0oribc1".to_string().parse().unwrap()),
+                    chain_id: "notibc0oribc1".to_string().parse().unwrap(),
                     ..default_params.clone()
                 },
                 want_pass: false,