+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [relayer-cli]
   - Implement command to query the channels associated with a connection ([#505])
   - JSON output for queries and txs ([#500])
+  - Added 'required' annotation for CLIs queries & txs; better error display ([#555])
   - Implement commands for channel close init and confirm ([#538])
   - Implement command to perform the handshake for a new channel ([#557])
   - Query all connections command ([#553])
@@ -71,6 +72,7 @@
 [#537]: https://github.com/informalsystems/ibc-rs/issues/537
 [#538]: https://github.com/informalsystems/ibc-rs/issues/538
 [#540]: https://github.com/informalsystems/ibc-rs/issues/540
+[#555]: https://github.com/informalsystems/ibc-rs/issues/555
 [#554]: https://github.com/informalsystems/ibc-rs/issues/554
 [#553]: https://github.com/informalsystems/ibc-rs/issues/553
 [#557]: https://github.com/informalsystems/ibc-rs/issues/557